feat: add explicit exports from package
