+++ .changeset/curly-numbers-shop.md
@@ -1,0 +1,5 @@
+---
+'vecteur': minor
+---
+
+feat: add explcit exports from package. use `vecteur/2d` or `vecteur/3d` instead of `vecteur` to import the correct module.
